moved last card in profile page to separate component with all logic to get data
